pingで疎通確認することによってエラーを回避？

+++ main.py
@@ -4,6 +4,8 @@ from wakeonlan import send_magic_packet
 
 import discord
 
+import pings
+
 # .envファイルの内容を読み込見込む
 load_dotenv()
 
@@ -11,6 +13,11 @@ intents = discord.Intents.default()
 intents.message_content = True
 
 client = discord.Client(intents=intents)
+
+# Ping
+p = pings.Ping()
+
+res = p.ping(os.environ["SSH"])
 
 
 # ログインしたとき
@@ -27,22 +34,31 @@ async def on_message(message):
 
     # サーバー起動
     if message.content.startswith("$サーバーを起動して"):
-        send_magic_packet(os.environ["WOL"])
-        await message.channel.send("サーバーを起動しました")
+        if res.is_reached():
+            await message.channel.send("サーバーは既に起動しています")
+        else:
+            send_magic_packet(os.environ["WOL"])
+            await message.channel.send("サーバーを起動しました")
 
     # サーバー停止
     if message.content.startswith("$サーバーを停止して"):
-        os.system(
-            f"sshpass -p {os.environ['RPASSWD']} ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null {os.environ['RUSERNAME']}@{os.environ['SSH']} 'echo {os.environ['RPASSWD']} | sudo -S shutdown -h now'"
-        )
-        await message.channel.send("サーバーを停止しました")
+        if res.is_reached():
+            os.system(
+                f"sshpass -p {os.environ['RPASSWD']} ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null {os.environ['RUSERNAME']}@{os.environ['SSH']} 'echo {os.environ['RPASSWD']} | sudo -S shutdown -h now'"
+            )
+            await message.channel.send("サーバーを停止しました")
+        else:
+            await message.channel.send("サーバーは既に閉じています。")
 
     # サーバー再起動
     if message.content.startswith("$サーバーを再起動して"):
-        os.system(
-            f"sshpass -p {os.environ['RPASSWD']} ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null {os.environ['RUSERNAME']}@{os.environ['SSH']} 'echo {os.environ['RPASSWD']} | sudo -S reboot'"
-        )
-        await message.channel.send("サーバーを再起動しました")
+        if res.is_reached():
+            os.system(
+                f"sshpass -p {os.environ['RPASSWD']} ssh -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null {os.environ['RUSERNAME']}@{os.environ['SSH']} 'echo {os.environ['RPASSWD']} | sudo -S reboot'"
+            )
+            await message.channel.send("サーバーを再起動しました")
+        else:
+            await message.channel.send("サーバーが閉じているので起動してください")
 
 
 # os.environを用いて環境変数を表示させます
